add signer state
