+++ signer/src/testing/context.rs
@@ -10,14 +10,14 @@ use blockstack_lib::{
     },
 };
 use clarity::types::chainstate::{StacksAddress, StacksBlockId};
-use tokio::sync::Mutex;
+use tokio::sync::{broadcast, Mutex};
 
 use crate::{
     bitcoin::{
         rpc::GetTxResponse, utxo::UnsignedTransaction, BitcoinInteract, MockBitcoinInteract,
     },
     config::Settings,
-    context::{Context, SignerContext},
+    context::{Context, SignerContext, SignerSignal, SignerState, TerminationHandle},
     emily_client::{EmilyInteract, MockEmilyInteract},
     error::Error,
     keys::PublicKey,
@@ -174,31 +174,35 @@ where
         self.inner.config()
     }
 
+    fn state(&self) -> &SignerState {
+        self.inner.state()
+    }
+
     fn get_signal_receiver(
         &self,
-    ) -> tokio::sync::broadcast::Receiver<crate::context::SignerSignal> {
+    ) -> broadcast::Receiver<SignerSignal> {
         self.inner.get_signal_receiver()
     }
 
-    fn get_signal_sender(&self) -> tokio::sync::broadcast::Sender<crate::context::SignerSignal> {
+    fn get_signal_sender(&self) -> broadcast::Sender<SignerSignal> {
         self.inner.get_signal_sender()
     }
 
-    fn signal(&self, signal: crate::context::SignerSignal) -> Result<(), Error> {
+    fn signal(&self, signal: SignerSignal) -> Result<(), Error> {
         self.inner.signal(signal)
     }
 
-    fn get_termination_handle(&self) -> crate::context::TerminationHandle {
+    fn get_termination_handle(&self) -> TerminationHandle {
         self.inner.get_termination_handle()
     }
 
-    fn get_storage(&self) -> impl crate::storage::DbRead + Clone + Sync + Send + 'static {
+    fn get_storage(&self) -> impl DbRead + Clone + Sync + Send + 'static {
         self.inner.get_storage()
     }
 
     fn get_storage_mut(
         &self,
-    ) -> impl crate::storage::DbRead + crate::storage::DbWrite + Clone + Sync + Send + 'static {
+    ) -> impl crate::storage::DbRead + DbWrite + Clone + Sync + Send + 'static {
         self.inner.get_storage_mut()
     }
